refactor - changed some classes names and created a new one to the constants / move methods from domain to service / changed domain from class to record / removed the CardDomain injection from the service class
