+++ src/main/java/com/jazztech/cardholder/infrastructure/persistence/mapper/CardHolderMapper.java
@@ -1,6 +1,6 @@
 package com.jazztech.cardholder.infrastructure.persistence.mapper;
 
-import com.jazztech.cardholder.domain.CardHolderDomain;
+import com.jazztech.cardholder.domain.entity.CardHolder;
 import com.jazztech.cardholder.infrastructure.persistence.entity.CardHolderEntity;
 import com.jazztech.cardholder.presentation.dto.CardHolderResponseDto;
 import java.util.List;
@@ -9,7 +9,7 @@ import org.mapstruct.Mapping;
 
 @Mapper(componentModel = "spring")
 public interface CardHolderMapper {
-    CardHolderEntity domainToEntity(CardHolderDomain cardHolderDomain);
+    CardHolderEntity domainToEntity(CardHolder cardHolder);
 
     @Mapping(source = "id", target = "cardHolderId")
     CardHolderResponseDto entityToDto(CardHolderEntity cardHolderEntity);
