+++ src/main/java/com/jazztech/cardholder/infrastructure/persistence/mapper/CardMapper.java
@@ -1,6 +1,6 @@
 package com.jazztech.cardholder.infrastructure.persistence.mapper;
 
-import com.jazztech.cardholder.domain.CardDomain;
+import com.jazztech.cardholder.domain.entity.Card;
 import com.jazztech.cardholder.infrastructure.persistence.entity.CardEntity;
 import com.jazztech.cardholder.presentation.dto.CardResponseDto;
 import org.mapstruct.Mapper;
@@ -8,7 +8,7 @@ import org.mapstruct.Mapping;
 
 @Mapper(componentModel = "spring")
 public interface CardMapper {
-    CardEntity domainToEntity(CardDomain cardDomain);
+    CardEntity domainToEntity(Card card);
 
     @Mapping(source = "id", target = "cardId")
     CardResponseDto entityToDto(CardEntity cardEntity);
