+++ src/main/java/com/jazztech/cardholder/service/constants/CardConstants.java
@@ -1,0 +1,20 @@
+package com.jazztech.cardholder.service.constants;
+
+public class CardConstants {
+    public static final Integer ROUND = 2;
+    public static final int AMERICAN_EXPRESS_MIN_RANGE = 34;
+    public static final int AMERICAN_EXPRESS_MAX_RANGE = 37;
+    public static final int MASTERCARD_MIN_RANGE = 51;
+    public static final int MASTERCARD_MAX_RANGE = 55;
+    public static final int RANDOM_NUM_RANGE = 10;
+    public static final int MODULUS = 10;
+    public static final int DIGIT_TEN = 10;
+    public static final int DIGIT_NINE = 9;
+    public static final int CVV_MAX_LENGTH = 3;
+    public static final int LUHN_DOUBLE_FACTOR = 2;
+    public static final int CARD_NUMBER_LENGTH = 15;
+    public static final int CARD_NUMBER_GROUP_SIZE = 4;
+
+    CardConstants() {
+    }
+}
